requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.9
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.0
django-reversion==2.0.13
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
jwt==0.5.3
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.0.7
odfpy==1.3.6
openpyxl==2.5.2
Pillow==5.1.0
pycparser==2.18
pycryptodome==3.6.1
PyJWT==1.6.1
PyMySQL==0.8.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
